Move validator extension point to model project - (Task #566)
