+++ de.dlr.sc.virsat.model/schema/de.dlr.sc.virsat.model.dvlm.validator.exsd
@@ -1,10 +1,10 @@
 <?xml version='1.0' encoding='UTF-8'?>
 <!-- Schema file written by PDE -->
-<schema targetNamespace="de.dlr.sc.virsat.build" xmlns="http://www.w3.org/2001/XMLSchema">
+<schema targetNamespace="de.dlr.sc.virsat.model" xmlns="http://www.w3.org/2001/XMLSchema">
 <annotation>
-      <appinfo>
-         <meta.schema plugin="de.dlr.sc.virsat.build" id="DvlmValidator" name="DVLM Validator"/>
-      </appinfo>
+      <appInfo>
+         <meta.schema plugin="de.dlr.sc.virsat.model" id="DvlmValidator" name="DVLM Validator"/>
+      </appInfo>
       <documentation>
          [Enter description of this extension point.]
       </documentation>
@@ -12,9 +12,9 @@
 
    <element name="extension">
       <annotation>
-         <appinfo>
+         <appInfo>
             <meta.element />
-         </appinfo>
+         </appInfo>
       </annotation>
       <complexType>
          <sequence minOccurs="1" maxOccurs="unbounded">
@@ -39,9 +39,9 @@
                <documentation>
                   
                </documentation>
-               <appinfo>
+               <appInfo>
                   <meta.attribute translatable="true"/>
-               </appinfo>
+               </appInfo>
             </annotation>
          </attribute>
       </complexType>
@@ -61,9 +61,9 @@
                <documentation>
                   
                </documentation>
-               <appinfo>
-                  <meta.attribute kind="java" basedOn=":de.dlr.sc.virsat.build.validator.external.IStructuralElementInstanceValidator"/>
-               </appinfo>
+               <appInfo>
+                  <meta.attribute kind="java" basedOn=":de.dlr.sc.virsat.model.dvlm.validator.IStructuralElementInstanceValidator"/>
+               </appInfo>
             </annotation>
          </attribute>
       </complexType>
@@ -83,9 +83,9 @@
                <documentation>
                   
                </documentation>
-               <appinfo>
-                  <meta.attribute kind="java" basedOn=":de.dlr.sc.virsat.build.validator.external.IRepositoryValidator"/>
-               </appinfo>
+               <appInfo>
+                  <meta.attribute kind="java" basedOn=":de.dlr.sc.virsat.model.dvlm.validator.IRepositoryValidator"/>
+               </appInfo>
             </annotation>
          </attribute>
       </complexType>
@@ -101,36 +101,36 @@
    </element>
 
    <annotation>
-      <appinfo>
+      <appInfo>
          <meta.section type="since"/>
-      </appinfo>
+      </appInfo>
       <documentation>
          [Enter the first release in which this extension point appears.]
       </documentation>
    </annotation>
 
    <annotation>
-      <appinfo>
+      <appInfo>
          <meta.section type="examples"/>
-      </appinfo>
+      </appInfo>
       <documentation>
          [Enter extension point usage example here.]
       </documentation>
    </annotation>
 
    <annotation>
-      <appinfo>
+      <appInfo>
          <meta.section type="apiinfo"/>
-      </appinfo>
+      </appInfo>
       <documentation>
          [Enter API information here.]
       </documentation>
    </annotation>
 
    <annotation>
-      <appinfo>
+      <appInfo>
          <meta.section type="implementation"/>
-      </appinfo>
+      </appInfo>
       <documentation>
          [Enter information about supplied implementation of this extension point.]
       </documentation>
