+++ de.dlr.sc.virsat.model/src/de/dlr/sc/virsat/model/dvlm/Activator.java
@@ -1,0 +1,73 @@
+/*******************************************************************************
+ * Copyright (c) 2008-2019 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+package de.dlr.sc.virsat.model.dvlm;
+
+
+import org.eclipse.core.runtime.Plugin;
+import org.osgi.framework.BundleContext;
+
+
+
+/**
+ * The plugin root class which is the entry point for the OSGi bundle activation
+ *
+ */
+public class Activator extends Plugin {
+
+	// The plug in ID
+	private static String pluginId;
+	// The shared instance
+	private static Activator plugin;
+
+	/**
+	 * The constructor
+	 */
+	public Activator() {
+	}
+
+	@Override
+	public void start(BundleContext context) throws Exception {
+		super.start(context);
+		plugin = this;
+		setPluginId(getDefault().getBundle().getSymbolicName());
+		
+	}
+
+	@Override
+	public void stop(BundleContext context) throws Exception {
+		plugin = null;
+		super.stop(context);
+	}
+
+	/**
+	 * Returns the shared instance
+	 *
+	 * @return the shared instance
+	 */
+	public static Activator getDefault() {
+		return plugin;
+	}
+	/**
+	 * Returns the plugin id
+	 *
+	 * @return pluginId
+	 */
+	public static String getPluginId() {
+		return pluginId;
+	}
+	/**
+	 * Sets the plugin id 
+	 *
+	 * @param pluginId the plugin id
+	 */
+	private static void setPluginId(String pluginId) {
+		Activator.pluginId = pluginId;
+	}
+}
